[CLI] Make CLI operate on a server-like architecture
